16 - Renderer

+++ OpenGL/src/Shader.cpp
@@ -116,7 +116,7 @@ void Shader::SetUniform1f(const std::string& name, float value)
 
 void Shader::SetUniform4f(const std::string& name, float v0, float v1, float v2, float v3)
 {
-	GLCall(glUniform4f(GetUniformLocation(name), v0, v1, v2, v3));
+	GLCall(glUniform4f(static_cast<GLint>(GetUniformLocation(name)), v0, v1, v2, v3));
 }
 
 int Shader::GetUniformLocation(const std::string& name)
